5/7

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const Product = require("../models/product");
 
+const fileHelper = require("../util/file");
+
 const { validationResult } = require("express-validator");
 
 exports.getAddProduct = (req, res, next) => {
@@ -142,6 +144,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
       if (image) {
         product.imageUrl = image.path.replace(/\\/g, "/");
+        fileHelper.deleteFile(product.imageUrl);
       }
 
       console.log("✅ Saving edited product:", {
@@ -177,10 +180,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log("DESTROYED PRODUCT");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => console.log(err));
+  Product.findById(prodId).then((product) => {
+    if (!product) {
+      return next(new Error("Product not found."));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    Product.deleteOne({ _id: prodId, userId: req.user._id })
+      .then(() => {
+        console.log("DESTROYED PRODUCT");
+        res.redirect("/admin/products");
+      })
+      .catch((err) => console.log(err));
+  });
 };
